Update to Fix issue #1 and Add more options
This is commit fixes #1 by adding graceful exits of errors and more HASH output types.

In this update you will now have to specify the domain and what SHA version you want back.

New Changes:
If you do not provide a domain it will connect to Localhost.
If not SHA Version is specified it will return nothing with an Exit code of 1.
If you specify more than one SHA hash version the Hashes are separated with `,` "CSV" list.

All of the CLI Flags that can be provided are as follows:
-sha1             Export the Sha1 Version of the Cert Fingerprint
-sha256         Export the Sha256 Version of the Cert Fingerprint
-sha512         Export the Sha512 Version of the Cert Fingerprint
-disableNth   Does not insert `:` into the output
-domain        The Domain to get the Fingerprint from.